If the lineNumberPanel tries to access an element that'd be out of bounds, it returns an empty list now.
This case usually only happens if the content of the diff changes quickly and the LineNumPanel tries to determine the lines when the diff is not done
loading into the document. The LineNumPanel should be triggered later on, at which point the lines will be calculated correctly. In the old case an
unnecessary exception was thrown