Changed all share/subscribeWith calls to replay/autoConnect, this combination has the same functionality but allows disposing the lower part of the chain
More info on the disposing the whole chain part: calling dispose on the observable returned by the subscribeWith call only disposes from the subscribeWith part onward,
the underlying chain is kept intact and there is no good way to get rid of it. With the replay and autoConnect version the functionality stays the same, and the
underlying chain may be disposed by calling dispose on the Disposables gathered by the second argument in the autoConnect call.
The implications of this are that quite a few additional classes now implement IDiscardable because they have some part of an observable chain that has to be disposed
(which just didn't happen before)