Changed the way that the treeModel is updated via the observable and how/when the before/after actions are triggered
The TreeModel itself no longer knows the Observable, instead a intermediary in the ObservableTreeUpdater is used. That ObservableTreeUpdater subscribes to the Observable
and tells the model to update itself with the new data each time the Observable changes. Before telling the model to update, the ObservableTreeUpdater executes all
actions that should happen before the model update. The After actions are still passed to the SwingWorker, who executes them after it is done with both the non-EDT and
EDT parts of its job