Dialogs can now have an unspecified number of buttons, not just the default OK and Cancel Buttons. Conflicts when unstashing should now be resolved properly
Having the option of using several different buttons means that the dialogs now have buttons that are more specific to the dialog. This should make it easier to do the
right thing for the user.
The "auto-stash" dialog now doesn't have the option "don't show again", this is so the users are more aware of changes they tow around

Signed-off-by: m.kaspera <m.kaspera@adito.de>